In this video we first changed the wallet creation to be with the user creation. Removing the Wallets insert from startup.
Then we changed the contacts to have userId, so we know who is the owner of these contacts.

We also used userId to filter wallet publication and contacts publication.

Later we fixed the transaction to have a destinationContactId instead of wallet.

We changed "Main account" in the home to be the email of the logged user and also made this email accessible in the first level of the object. We also included this new property in the default publish fields.

And last we fixed the home page to be empty when no user is logged in.